Create LICENSE
I think MIT License would be a good fit for this project. The simplicity and clarity of the MIT License were also factors. It is a concise and easily understandable license, ensuring that users can quickly comprehend and adhere to its terms. MIT license promotes collaboration and fosters a vibrant open-source community. In this project, there is a possibility to use a pre-trained model, the modifications and improvements that are made to the model in this project should be made available to other developers in the community.